+++ cynic-codegen/src/suggestions/mod.rs
@@ -5,8 +5,7 @@ pub fn guess_field<'a>(
     candidates: impl Iterator<Item = &'a str>,
     field_name: &str,
 ) -> Option<&'a str> {
-    return candidates
-        .min_by_key(|candidate| hamming(candidate, field_name).unwrap_or(usize::max_value()));
+    candidates.min_by_key(|candidate| hamming(candidate, field_name).unwrap_or(usize::max_value()))
 }
 
 pub fn format_guess(guess_field: Option<&str>) -> String {