+++ tests/ui-tests/tests/cases/fragment-guess-validation.rs
@@ -1,11 +1,11 @@
 fn main() {}
 
 #[cynic::query_module(
-schema_path = r#"./../../../schemas/starwars.schema.graphql"#,
-query_module = "query_dsl",
+    schema_path = r#"./../../../schemas/starwars.schema.graphql"#,
+    query_module = "query_dsl"
 )]
 mod queries {
-    use super::{query_dsl, types::*};
+    use super::query_dsl;
 
     #[derive(cynic::QueryFragment, Debug)]
     #[cynic(graphql_type = "Root")]
@@ -29,14 +29,6 @@ mod queries {
     }
 }
 
-#[cynic::query_module(
-schema_path = r#"./../../../schemas/starwars.schema.graphql"#,
-query_module = "query_dsl",
-)]
-mod types {}
-
 mod query_dsl {
-    use super::types::*;
-
     cynic::query_dsl!(r#"./../../../schemas/starwars.schema.graphql"#);
 }